feat(report): add Parquet report generation via unified provider and versioned metadata (#33)
## Overview
- Standardize report generation behind a single-call provider interface.
- Enable generating the same HTML report from Parquet outputs.
- Keep Parquet write-time ordering unchanged and leverage versioned
metadata to minimize reads.

## Scope
- Add a unified provider interface that returns a complete `ReportData`.
- Support report generation from Parquet batches with metadata
versioning (v1) and strict validation.

## Design

### 1) Provider Abstraction
- A single-call provider interface returning `ReportData`, shared by
ClickHouse and Parquet paths.

### 2) Parquet File Metadata (msgpack+base64)
- Keys:
  - `rdbinsight.meta.version = "1"`
  - `rdbinsight.meta.encoding = "msgpack"`
  - `rdbinsight.meta.summary.b64 = "<base64>"`
- Summary payload (per file):
  - `cluster`, `batch_unix_nanos`, `instance`
  - `total_key_count`, `total_size_bytes`
  - `per_db[]`, `per_type[]`
  - `top_keys_full[]` (full `Record`, up to 100)
  - optional `slots` (`codis_slots[]`, `redis_slots[]`)
  - `per_db_row_groups` (`Map<db, Vec<row_group_idx>>`)

### 3) Read-Time Plan (from Parquet)
- Validate metadata version/encoding; decode summary and synthesize:
  - db/type/instance aggregates, totals, full Top100, slot skew
- Build a global key-ordered stream:
  - multi-cursor substreams (per file, per db) via `per_db_row_groups`
  - min-heap on key; group duplicate keys before advancing
- Dynamic prefix discovery:
  - single pass with fixed threshold (>= 1% total size)
  - scan only projected columns `{key, rdb_size}`
- Output JSON compatible with the existing HTML template.

## CLI
- Add `report from-parquet` to target cluster/batch and emit a
single-file HTML report.

## Compatibility & Constraints
- Only Parquet files produced by this tool with metadata version `"1"`
are supported; mismatches fail fast.
- Parquet write-time ordering remains `(cluster, batch, instance, db,
key)`.

## Out of Scope
- General fallback via external OLAP engines (e.g., DuckDB).
- Low-level implementation details beyond the provider abstraction.

+++ src/output/parquet/schema.rs
@@ -11,21 +11,21 @@ pub fn create_redis_record_schema() -> Schema {
             false,
         ),
         Field::new("instance", DataType::Utf8, false),
-        Field::new("db", DataType::Int64, false),
+        Field::new("db", DataType::UInt64, false),
         Field::new("key", DataType::Binary, false),
         Field::new("type", DataType::Utf8, false),
-        Field::new("member_count", DataType::Int64, false),
-        Field::new("rdb_size", DataType::Int64, false),
+        Field::new("member_count", DataType::UInt64, false),
+        Field::new("rdb_size", DataType::UInt64, false),
         Field::new("encoding", DataType::Utf8, false),
         Field::new(
             "expire_at",
             DataType::Timestamp(TimeUnit::Millisecond, Some("UTC".into())),
             true, // nullable
         ),
-        Field::new("idle_seconds", DataType::Int64, true), // nullable
-        Field::new("freq", DataType::Int32, true),         // nullable
-        Field::new("codis_slot", DataType::Int32, true),   // nullable
-        Field::new("redis_slot", DataType::Int32, true),   // nullable
+        Field::new("idle_seconds", DataType::UInt64, true), // nullable
+        Field::new("freq", DataType::UInt8, true),          // nullable
+        Field::new("codis_slot", DataType::UInt16, true),   // nullable
+        Field::new("redis_slot", DataType::UInt16, true),   // nullable
     ])
 }
